requirements.txt
torch
torchvision
matplotlib
tqdm
numpy
Pillow  
seaborn
scikit-learn
